Change uint16_t cast to uintptr_t cast
GCC gave a lot of warnings for this, but I initially did not want to
change the TI provided library/header. But it got difficult to find my
warnings with the amount of warnings it created.

+++ nvm-controller/MSP430FR5xx_6xx/inc/hw_memmap.h
@@ -68,10 +68,10 @@
 //
 //*****************************************************************************
 #define HWREG32(x)                                                              \
-        (*((volatile uint32_t *)((uint16_t)x)))
+        (*((volatile uint32_t *)((uintptr_t)x)))
 #define HWREG16(x)                                                             \
-        (*((volatile uint16_t *)((uint16_t)x)))
+        (*((volatile uint16_t *)((uintptr_t)x)))
 #define HWREG8(x)                                                             \
-        (*((volatile uint8_t *)((uint16_t)x)))
+        (*((volatile uint8_t *)((uintptr_t)x)))
 
 #endif // #ifndef __HW_MEMMAP__
